converted to big decimal ang mga double variable amounts, removed computeDays kasi get days nalang from timecards, changes in updatePayroll

+++ src/main/java/Entity/Employee.java
@@ -1,5 +1,7 @@
 package Entity;
 
+import java.math.BigDecimal;
+
 public class Employee {
     private int employee_id;
     private String last_name;
@@ -8,7 +10,7 @@ public class Employee {
     private String tin_number;
     private String philhealth_number;
     private String pagibig_number;
-    private double pay_rate;
+    private BigDecimal pay_rate;
     private String employment_status;
     private String department;
     private String shift_start;
@@ -16,7 +18,7 @@ public class Employee {
 
     // Constructor
 
-    public Employee(int employee_id, String last_name, String first_name, String middle_name, String tin_number, String philhealth_number, String pagibig_number, double pay_rate, String employment_status, String department, String shift_start, String shift_end ){
+    public Employee(int employee_id, String last_name, String first_name, String middle_name, String tin_number, String philhealth_number, String pagibig_number, BigDecimal pay_rate, String employment_status, String department, String shift_start, String shift_end ){
         this.employee_id = employee_id;
         this.last_name = last_name;
         this.first_name = first_name;
@@ -31,7 +33,7 @@ public class Employee {
         this.shift_end = shift_end;
     }
 
-    public Employee(int employee_id, String last_name, String first_name, double pay_rate, String department){
+    public Employee(int employee_id, String last_name, String first_name, BigDecimal pay_rate, String department){
             this.employee_id = employee_id;
             this.last_name = last_name;
             this.first_name = first_name;
@@ -50,7 +52,7 @@ public class Employee {
     public  String getTin_number(){return tin_number;}
     public String getPhilhealth_number(){return philhealth_number;}
     public String getPagibig_number() {return pagibig_number;}
-    public double getPay_rate(){return pay_rate;}
+    public BigDecimal getPay_rate(){return pay_rate;}
     public String getEmployment_status(){return employment_status;}
     public String getDepartment(){return department;}
     public String getShift_start(){return shift_start;}
